Create README.md
In this python project video, you will learn how to check internet speed on your device with the help of Python.

With our free project tutorial series, practice beginner-level projects and become a pro. You can also get to know about the internet speed. You can check upload speed and download speed on your devices.

Python internet speed test code is used after installing a library in it. The library that is needed for the Python internet speed project is the speed test. You can change the font type, background color, borders, and sizes with the help of this project.

Python internet speed check will be the first project toward going pro with the language. Our trainer will take you through the basics of this project by giving proper examples.

You can install the library with the help of the pip command. Python is used widely in programmers, and learning this will help you boost your career. Internet speed test with Python is one of the many projects you can watch on YouTube for free.